set format onRunQuery new button
